+++ gms-server/src/main/java/org/gms/model/dto/GiveResourceReqDTO.java
@@ -35,8 +35,8 @@ public class GiveResourceReqDTO {
     private Short speed;
     private Short jump;
     private Byte upgradeSlot;
-    private Byte level;
-    private Byte itemLevel;
+    private Short level;
+    private Short itemLevel;
     private Long expire;
     private String owner;
     private Short flag;
